extract the source catalog into connectors/source-catalog.ts
